Implement basic keyboard shortcuts for selecting cards
