+++ package.json
@@ -17,6 +17,7 @@
     "react": "^16.13.1",
     "react-chartjs-2": "^2.9.0",
     "react-dom": "^16.13.1",
+    "react-hotkeys": "^2.0.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "^3.4.1",
     "react-spring": "^8.0.27"
